Add files via upload
These are the basic files my group and I created for my CSC270 - Applied Databases final project. The idea was that we would allow students at The College of Idaho to log in and share information about classes and eventually add functionality for exchanging books. We also implemented a page that allowed students to search a databases of classes offered to assist in the creation of their "4 Year Schedule".